+++ cloudformation.go
@@ -2,10 +2,20 @@ package cit
 
 import (
 	"context"
-
+	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/cloudformation"
 )
 
+var CfnClient *cloudformation.Client
+
+
+func init(){
+	cfg, err := config.LoadDefaultConfig(context.TODO())
+	if err != nil {
+		panic("configuration error, " + err.Error())
+	}
+	CfnClient  = cloudformation.NewFromConfig(cfg)
+}
 //go:generate moq -out cloudformation_moq_test.go . CloudFormationInterface
 
 type CloudFormationInterface interface {