+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.1.0] - 2021-06-02
+## [0.1.1] - 2021-06-02
 ### Added
-- Get physical ID from Contstruct ID
+- Get VPC type struct directly with GetVPC
+- Get physical ID from Construct ID